Replace nock integration test by fetch-mock
