+++ __tests__/projects.js
@@ -15,126 +15,6 @@ const rpm = new RepositoryProjectsManager({
   repositoryName: 'example-repository',
 });
 
-const nockHTTPRequestsForAssigningSinglePR = () => {
-  nock('https://api.github.com')
-    .post('/graphql', (body) => /projectsV2.first:/.test(body.query))
-    .reply(
-      200,
-      {
-        data: {
-          repository: {
-            name: 'example-repository',
-            id: 'R_0000000001',
-            projectsV2: {
-              nodes: [
-                {
-                  id: 'PVT_0000000000000001',
-                  title: 'layer-100/bar',
-                  number: 1099,
-                  fields: {
-                    nodes: [
-                      {},
-                      {},
-                      {
-                        id: 'PVTSSF_00000000000000000000001',
-                        name: 'Status',
-                        options: [
-                          {
-                            id: '00000001',
-                            name: 'Todo',
-                          },
-                          {
-                            id: '00000002',
-                            name: 'In Progress',
-                          },
-                          {
-                            id: '00000003',
-                            name: 'Done',
-                          },
-                        ],
-                      },
-                      {},
-                      {},
-                    ],
-                  },
-                },
-              ],
-              pageInfo: {
-                hasNextPage: false,
-                endCursor: 'Nw',
-              },
-            },
-          },
-        },
-      },
-    )
-    .post('/graphql', (body) => /on PullRequest/.test(body.query))
-    .reply(
-      200,
-      {
-        data: {
-          node:
-             {
-               number: 74,
-               projectsV2: {
-                 nodes: [],
-                 pageInfo: {
-                   hasNextPage: false,
-                   endCursor: 'MQ',
-                 },
-               },
-             },
-        },
-      },
-    )
-    .post('/graphql', (body) => /addProjectV2ItemById/.test(body.query))
-    .reply(
-      200,
-      {
-        data: {
-          addProjectV2ItemById: {
-            item: {
-              id: 'PVTI_0000000000000001',
-            },
-          },
-        },
-      },
-    )
-    .post('/graphql', (body) => /updateProjectV2ItemFieldValue/.test(body.query))
-    .reply(
-      200,
-      {
-        data: {
-          updateProjectV2ItemFieldValue: { projectV2Item: { id: 'PVTI_0000000000000001' } },
-        },
-      },
-    )
-    .post('/graphql', (body) => /on PullRequest/.test(body.query))
-    .reply(
-      200,
-      {
-        data: {
-          node:
-             {
-               number: 74,
-               projectsV2: {
-                 nodes: [
-                   {
-                     id: 'PVT_0000000000000001',
-                     title: 'layer-100/bar',
-                   },
-                 ],
-                 pageInfo: {
-                   hasNextPage: false,
-                   endCursor: 'MQ',
-                 },
-               },
-             },
-        },
-      },
-    );
-};
-
 const nockHTTPRequestsForAlreadyAssignedProject = () => {
   nock('https://api.github.com')
     .post('/graphql', (body) => /projectsV2.first:/.test(body.query))
@@ -423,18 +303,6 @@ describe('RepositoryProjectsManager.assing() posts requests to the API', () => {
 
   afterEach(() => {
     nock.cleanAll();
-  });
-
-  test('when the PR is not assigned to a project yet', async () => {
-    const titles = ['layer-100/bar'];
-
-    const pullRequestNumber = 'PR_0000000000000001';
-    nockHTTPRequestsForAssigningSinglePR();
-
-    const outputProjects = await rpm.assign(pullRequestNumber, titles);
-    const outputTitles = outputProjects.map((p) => p.title);
-
-    expect(outputTitles).toEqual(titles);
   });
 
   // setup and requests are identical fore example for assigning the PR to a new project because
